+++ src/homepage2/App.js
@@ -1,28 +1,24 @@
 // @flow
 
 import React from "react";
-import {Router} from "react-router";
-import type {History /* actually `any` */} from "history";
 
-import {createRoutes} from "./createRoutes";
-import {type RouteData, resolveTitleFromPath} from "./routeData";
+async function loadAndReport(path) {
+  const response = await fetch(path);
+  if (!response.ok) {
+    console.error(path, response);
+  }
+  const json = await response.json();
+  console.log(path, json);
+}
 
-export default class App extends React.Component<{|
-  +routeData: RouteData,
-  +history: History,
-|}> {
+export default class App extends React.Component<{||}> {
+  async componentDidMount() {
+    loadAndReport("sourcecred.json");
+    loadAndReport("output/credResult.json");
+    loadAndReport("config/sourcecred/discourse/config.json");
+  }
+
   render() {
-    const {routeData, history} = this.props;
-    return (
-      <Router
-        history={history}
-        routes={createRoutes(routeData)}
-        onUpdate={function () {
-          const router = this;
-          const path: string = router.state.location.pathname;
-          document.title = resolveTitleFromPath(routeData, path);
-        }}
-      />
-    );
+    return <h1>Under Construction</h1>;
   }
 }
